style: fix format-check + clang-tidy findings
- test/sql/otlp_serve.test: blank line after header block (SQL test
  formatter, clang-format-11 / format.py --check).
- otlp_arrow.cpp: avoid int-multiplication widening in the FixedSizeBinary
  hex loop; cast idx_t offset to int64_t before adding to ArrowArray.offset
  (bugprone-implicit-widening-of-multiplication-result, bugprone-narrowing-
  conversions).
- otlp_server.cpp / otlp_server.hpp: take HttpOtlpServer config by const ref
  (performance-unnecessary-value-param); emplace_back the listener thread
  (modernize-use-emplace).

Verified locally with the CI toolchain: clang-format 11.0.1 + black 24 via
format.py --all --check (clean), and clang-tidy with duckdb/.clang-tidy
(clean). Build + otlp SQL suite green.

+++ src/otlp_arrow.cpp
@@ -326,8 +326,8 @@ void CopyArrowToDuckDB(const ArrowArray &array, const ArrowSchema &schema, Vecto
 			}
 			const uint8_t *row = bytes + array_idx * static_cast<size_t>(width);
 			for (int b = 0; b < width; b++) {
-				buf[2 * b] = hex[row[b] >> 4];
-				buf[2 * b + 1] = hex[row[b] & 0x0F];
+				buf[2 * static_cast<size_t>(b)] = hex[row[b] >> 4];
+				buf[2 * static_cast<size_t>(b) + 1] = hex[row[b] & 0x0F];
 			}
 			string_data[i] = StringVector::AddString(output, buf.data(), buf.size());
 		}
@@ -410,7 +410,7 @@ void CopyArrowStructToDataChunk(const ArrowArray &array, const ArrowSchema &sche
 			throw IOException("Invalid Arrow batch: column %llu schema is null", static_cast<uint64_t>(col_idx));
 		}
 		ArrowArray col_view = *col_array;
-		col_view.offset = col_array->offset + offset;
+		col_view.offset = col_array->offset + static_cast<int64_t>(offset);
 		CopyArrowToDuckDB(col_view, *col_schema, output.data[col_idx], count);
 	}
 }
